requirements.txt
Django==1.5.1
Pillow==2.1.0
reportlab==2.7
requests==1.2.3
wsgiref==0.1.2
